+++ bundles/mapping/mapmodule/mapmodule.ol2.js
@@ -282,7 +282,7 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
          * Moves the map to the given position.
          * NOTE! Doesn't send an event if zoom level is not changed.
          * Call notifyMoveEnd() afterwards to notify other components about changed state.
-         * @param {OpenLayers.LonLat} or {Array} lonlat coordinates to move the map to
+         * @param {Number[] | Object} or {Array} lonlat coordinates to move the map to
          * @param {Number} zoomAdjust relative change to the zoom level f.ex -1 (optional)
          * @param {Boolean} pIsDragging true if the user is dragging the map to a new location currently (optional)
          */
@@ -739,35 +739,31 @@ Oskari.clazz.define('Oskari.mapframework.ui.module.common.MapModule',
             mapVO.setBbox(this._map.calculateBounds());
         },
 
+
         /**
          * @method transformCoordinates
-         * Deprecated
-         * Transforms coordinates from given projection to the maps projectino.
-         * @param {OpenLayers.LonLat} pLonlat
+         * Transforms coordinates from given projection to the maps projection.
+         * @param {Object} pLonlat object with lon and lat keys
          * @param {String} srs projection for given lonlat params like "EPSG:4326"
-         * @return {OpenLayers.LonLat} transformed coordinates
+         * @return {Object} transformed coordinates as object with lon and lat keys
          */
         transformCoordinates: function (pLonlat, srs) {
-            this.getSandbox().printWarn(
-                'transformCoordinates is deprecated. Use _transformCoordinates instead if called from plugin. Otherwise, use Requests instead.'
-            );
+            if(this.getProjection() === srs) {
+                return pLonlat;
+            }
+            var isProjectionDefined = Proj4js.defs[srsName];
+            if (!isProjectionDefined) {
+                throw 'SrsName not supported! Provide Proj4js.def for ' + srs;
+            }
+            var tmp = new OpenLayers.LonLat(pLonlat.lon, pLonlat.lat);
+            var transformed = tmp.transform(new OpenLayers.Projection(srs), this.getMap().getProjectionObject());
 
-            return this._transformCoordinates(pLonlat, srs);
+            return {
+                lon : transformed.lon,
+                lat : transformed.lat
+            };
         },
 
-        /**
-         * @method _transformCoordinates
-         * Transforms coordinates from given projection to the maps projectino.
-         * @param {OpenLayers.LonLat} pLonlat
-         * @param {String} srs projection for given lonlat params like "EPSG:4326"
-         * @return {OpenLayers.LonLat} transformed coordinates
-         */
-        _transformCoordinates: function (pLonlat, srs) {
-            return pLonlat.transform(
-                new OpenLayers.Projection(srs),
-                this.getMap().getProjectionObject()
-            );
-        },
         /**
          * @property eventHandlers
          * @static